Examples of Painter's Alg and Reverse Painter's Alg in C++
Skeleton code is from the Painters1 file that just creates a window. The display() and updateBuffer() is new code I've added based on Tom's work and research on the glDrawPixels method. Those methods show the minor differences between the original algorithm and reverse algorithm. I didn't do this in Visual Studio so there might be a couple errors. I think this decently exhibits the general concept of a z-buffer, but feel free to make any necessary fixes/suggestions.